feat: add 'deltaTime' for time control and other modifications.

+++ src/StardustSandbox/Enums/InputSystem/MouseButton.cs
@@ -1,6 +1,6 @@
 ﻿namespace StardustSandbox.Enums.InputSystem
 {
-    internal enum MouseButton : short
+    internal enum MouseButton : sbyte
     {
         Left = -1,
         Middle = 0,
